+++ web/renderer/components/pageComponents/ConnectionsPage/NewConnection/context/config.tsx
@@ -10,13 +10,15 @@ import {
 } from "@gen/graphql-types";
 import useMutation from "@hooks/useMutation";
 import { useRouter } from "next/router";
-import { ReactNode, SyntheticEvent, useEffect, useMemo } from "react";
+import { ReactNode, SyntheticEvent, useEffect, useMemo, useState } from "react";
 import { maybeDatabase } from "@lib/urls";
 import { ConfigContextType, defaultState, getDefaultState } from "./state";
 import { getConnectionUrl } from "./utils";
 
 export const ConfigContext =
   createCustomContext<ConfigContextType>("ConfigContext");
+
+const forElectron = process.env.NEXT_PUBLIC_FOR_ELECTRON === "true";
 
 type Props = {
   children: ReactNode;
@@ -31,6 +33,9 @@ export function ConfigProvider({ children }: Props) {
   });
 
   const connectionsRes = useStoredConnectionsQuery();
+  const [err, setErr] = useState<Error | undefined>(
+    res.err || connectionsRes.error,
+  );
 
   useEffectOnMount(() => {
     const isDocker = window.location.origin === "http://localhost:3000";
@@ -45,6 +50,13 @@ export function ConfigProvider({ children }: Props) {
       setState({ showAdvancedSettings: true });
     }
   }, [res.err]);
+
+  useEffect(() => {
+    if (!forElectron) return;
+    window.ipc.getDoltServerError(async (msg: string) => {
+      setErr(Error(msg));
+    });
+  }, []);
 
   const clearState = () => {
     setState(defaultState);
@@ -81,18 +93,49 @@ export function ConfigProvider({ children }: Props) {
     }
   };
 
+  const onStartDoltServer = async (e: SyntheticEvent) => {
+    e.preventDefault();
+    setState({ loading: true });
+    try {
+      const result = await window.ipc.invoke(
+        "start-dolt-server",
+        state.name,
+        state.port,
+        true,
+      );
+
+      if (result !== "success") {
+        setErr(Error(result));
+        return;
+      }
+      await onSubmit(e);
+    } catch (error) {
+      setErr(Error(` ${error}`));
+    } finally {
+      setState({ loading: false });
+    }
+  };
+
   const value = useMemo(() => {
     return {
       state,
       setState,
       onSubmit,
-      error: res.err,
-      setErr: res.setErr,
-
+      error: err,
+      setErr,
       clearState,
       storedConnections: connectionsRes.data?.storedConnections,
+      onStartDoltServer,
     };
-  }, [state, setState, onSubmit, res.err, clearState, connectionsRes]);
+  }, [
+    state,
+    setState,
+    onSubmit,
+    res.err,
+    clearState,
+    connectionsRes,
+    onStartDoltServer,
+  ]);
 
   return (
     <ConfigContext.Provider value={value}>{children}</ConfigContext.Provider>
